[UI][E] Add project creation button to resource negotiation wizard
Adds a button to create a new project to the wizard used to accept
incoming resource negotiations. This button can be used by the user to
create and configure new projects that can then subsequently be used to
represent the shared reference points in the local workspace.

This option is necessary to ensure that the wizard is always
completable, even in cases where the workspace is empty when the wizard
is started.

As the button is part of the external wizard and not the selection tab
itself, creating a new project through the button does not update the
currently selected tab. This has to be done manually by the user.